+++ src/dune_rules/coq_stanza.ml
@@ -14,13 +14,13 @@ let coq_syntax =
 
 module Coqpp = struct
   type t =
-    { modules : string list
+    { modules : Ordered_set_lang.t
     ; loc : Loc.t
     }
 
   let decode =
     fields
-      (let+ modules = field "modules" (repeat string)
+      (let+ modules = Ordered_set_lang.field "modules"
        and+ loc = loc in
        { modules; loc })